Allow test_helper to build Command with script
Since `std::process::Command` cannot run more than one command, creating
scripts for the purposes of testing becomes a real chore. The
alternative is to pass the shell script as the `-c` argument to shell to
get around this limitation.

This commit adds a `test_helper` function that encapsulates this
behavior.